refactor(jdbc): 패키지 정리
