+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -3,7 +3,7 @@ package com.techcourse.dao;
 import com.techcourse.domain.User;
 import java.util.List;
 import nextstep.jdbc.JdbcTemplate;
-import nextstep.jdbc.RowMapper;
+import nextstep.jdbc.resultset.RowMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
