+++ jdbc/src/main/java/nextstep/jdbc/resultset/ResultSetExecutor.java
@@ -1,7 +1,8 @@
-package nextstep.jdbc;
+package nextstep.jdbc.resultset;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import nextstep.jdbc.exception.DataAccessException;
 
 @FunctionalInterface
 public interface ResultSetExecutor<T> {
