+++ jdbc/src/main/java/nextstep/jdbc/resultset/RowMapper.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.resultset;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
